user -> userName
Changed user to userName when passing it to functions.
Added some visual feedback to the console so you know if it's even seeing messages.
Added the bot's name and discriminator so you know what bot it's using.
Updated clean_user_message so it cleans the bot's actual name & discriminator
Send the channel.typing() as another visual indicator the bot is intending to reply.
Pass the full user (message.author) to the queue. [for future use]